requirements.txt
requests>=2.32.0
pycryptodome>=3.20.0

